Update: README in both client, server apps and database configuration
