+++ q-file-share-server/app/models/dto.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+
 class LoginRequest(BaseModel):
     email: str
     password: str
